+++ build-config/fragments/common.js
@@ -26,16 +26,6 @@ module.exports = {
         rules: [
             {
                 test: /\.js$/,
-                enforce: 'pre',
-                exclude: /node_modules/,
-                use: [
-                    {
-                        loader: 'eslint-loader'
-                    }
-                ]
-            },
-            {
-                test: /\.js$/,
                 exclude: /node_modules/,
                 use: [
                     {