+++ karma.conf.js
@@ -15,13 +15,15 @@ var chromeFlags = [
     '--no-first-run',
     '--noerrdialogs',
     '--no-default-browser-check',
-    '--autoplay-policy=no-user-gesture-required',
     '--user-data-dir=.chrome',
     '--disable-translate',
     '--disable-extensions',
     '--disable-infobars',
     '--ignore-certificate-errors',
-    '--allow-insecure-localhost'
+    '--allow-insecure-localhost',
+    '--autoplay-policy=no-user-gesture-required',
+    // see https://developers.google.com/web/updates/2017/09/autoplay-policy-changes#webaudio
+    '--disable-features=PreloadMediaEngagementData,AutoplayIgnoreWebAudio,MediaEngagementBypassAutoplayPolicies'
 ];
 
 module.exports = function(config) {
@@ -46,13 +48,15 @@ module.exports = function(config) {
             'spec/plugin-api.spec.js',
             'spec/util.spec.js',
             'spec/wavesurfer.spec.js',
-            'spec/peakcache.spec.js'
+            'spec/peakcache.spec.js',
+            'spec/mediaelement.spec.js'
         ],
         preprocessors: {
             'spec/plugin-api.spec.js': ['webpack'],
             'spec/util.spec.js': ['webpack'],
             'spec/wavesurfer.spec.js': ['webpack'],
             'spec/peakcache.spec.js': ['webpack'],
+            'spec/mediaelement.spec.js': ['webpack'],
 
             // source files, that you want to generate coverage for
             // do not include tests or libraries
@@ -71,7 +75,7 @@ module.exports = function(config) {
             'karma-coveralls',
             'karma-verbose-reporter'
         ],
-        browsers: ['Chrome_dev', 'FirefoxHeadless'],
+        browsers: ['Chrome_ci', 'FirefoxHeadless'],
         captureConsole: true,
         colors: true,
         reporters: ['verbose', 'progress', 'coverage'],