+++ src/mediaelement.js
@@ -64,10 +64,7 @@ export default class MediaElement extends WebAudio {
             this.fireEvent('audioprocess', this.getCurrentTime());
 
             // Call again in the next frame
-            const requestAnimationFrame =
-                window.requestAnimationFrame ||
-                window.webkitRequestAnimationFrame;
-            requestAnimationFrame(onAudioProcess);
+            util.frame(onAudioProcess)();
         };
 
         this.on('play', onAudioProcess);
@@ -119,8 +116,8 @@ export default class MediaElement extends WebAudio {
     }
 
     /**
-     * Private method called by both load (from url)
-     * and loadElt (existing media element).
+     * Private method called by both `load` (from url)
+     * and `loadElt` (existing media element) methods.
      *
      * @param {HTMLMediaElement} media HTML5 Audio or Video element
      * @param {number[]|number[][]} peaks Array of peak data
@@ -148,7 +145,7 @@ export default class MediaElement extends WebAudio {
             this.fireEvent('finish');
         });
 
-        // Listen to and relay play and pause events to enable
+        // Listen to and relay play, pause and seeked events to enable
         // playback control from the external media element
         media.addEventListener('play', () => {
             this.fireEvent('play');
@@ -156,6 +153,10 @@ export default class MediaElement extends WebAudio {
 
         media.addEventListener('pause', () => {
             this.fireEvent('pause');
+        });
+
+        media.addEventListener('seeked', event => {
+            this.fireEvent('seek');
         });
 
         this.media = media;